+++ Source/Utility/Time/Clock.cpp
@@ -4,9 +4,9 @@ using namespace Grim;
 void Clock::initialize()
 {
 	QueryPerformanceFrequency(&m_LargeInt);
-	m_Frequency = static_cast<float32>(m_LargeInt.QuadPart);
+	m_Frequency = m_LargeInt.QuadPart;
 	QueryPerformanceCounter(&m_LargeInt);
-	m_CounterStart = static_cast<float32>(m_LargeInt.QuadPart);
+	m_CounterStart = m_LargeInt.QuadPart;
 }
 
 void Clock::terminate()
